+++ templates/admin-panel/template/client/src/initializers/main.js
@@ -3,6 +3,7 @@ import hrefto from 'sham-ui-router/href-to';
 import { storage as appState } from '../storages/app';
 import Session from '../services/session';
 import Store from '../services/store';
+import Title from '../services/title';
 import startRouter from './routes';
 import App from '../components/App.sfc';
 
@@ -11,6 +12,7 @@ export default function() {
     // Create services
     const session = new Session();
     const store = new Store();
+    new Title();
 
     // Mount root component
     new App( {