+++ templates/admin-panel/template/client/src/services/title.js
@@ -1,0 +1,16 @@
+import { DI } from 'sham-ui';
+
+
+export default class Title {
+    constructor() {
+        DI.bind( 'title', this );
+    }
+
+    /**
+     * Set new document title
+     * @param {string} newTitle
+     */
+    change( newTitle ) {
+        document.title = `{{logoText}} | ${newTitle}`;
+    }
+}